+++ src/sora/sora_client.cpp
@@ -685,20 +685,28 @@ void SoraClient::OnMessage(
     return;
   }
 
-  // zakuro の場合、DataChannel メッセージは Fake データだけのはずで、
-  // その場合は先頭 16 バイトに特定のデータが入っている。
+  // zakuro の DataChannel メッセージだった場合は先頭6バイトが "ZAKURO" になっていて、
+  // その場合はその後 16 バイトに特定のデータが入っている。
   if (label[0] == '#') {
-    if (data.size() < 16) {
+    if (data.size() < 22) {
       return;
     }
+    if (std::memcmp(data.c_str(), "ZAKURO", 6) != 0) {
+      return;
+    }
+    const char* p = data.c_str() + 6;
+
     uint64_t time = 0;
     for (int i = 0; i < 8; i++) {
-      time |= ((uint64_t)data[i] & 0xff) << ((7 - i) * 8);
+      time |= ((uint64_t)p[i] & 0xff) << ((7 - i) * 8);
     }
+    p += 8;
+
     uint64_t counter = 0;
     for (int i = 0; i < 8; i++) {
-      counter |= ((uint64_t)data[i + 8] & 0xff) << ((7 - i) * 8);
+      counter |= ((uint64_t)p[i] & 0xff) << ((7 - i) * 8);
     }
+
     RTC_LOG(LS_INFO) << "Recv DataChannel unixtime(us)=" << time
                      << " counter=" << counter;
     return;